Retrieval REST methods implemented, apparently
All designed REST methods and their endpoints have been implemented. Even so, I haven't checked the correct functionality of each of them (due to lack of data and my desire to sleep). Some work still has to be done, but I can say that this service is totally functional.

+++ AppointmentService/src/main/java/interfaces/IAppointmentService.java
@@ -22,7 +22,7 @@ public interface IAppointmentService {
     public CancelAppointmentResponse cancelAppointment(CancelAppointmentRequest request);
     // Queries
     public BookingDTO getBookingById(Integer id);
-    public BookingDTO getBookingByReferenceNumber(Integer reference);
+    public BookingDTO getBookingByReferenceNumber(Long reference);
     public List<BookingDTO> getAllBookings();
     public List<BookingDTO> getBookingsByPatient(Integer patientId);
     public List<BookingDTO> getBookingsByPatientStatus(Integer patientId, BookingStatus status);
